LayoutTests/imported/w3c/web-platform-tests/dom/events/event-global-expected.txt


PASS event exists on window, which is initially set to undefined
PASS window.event is only defined during dispatch
PASS window.event is undefined if the target is in a shadow tree (event dispatched outside shadow tree)
PASS window.event is undefined if the target is in a shadow tree (event dispatched inside shadow tree)
PASS window.event is set to the current event during dispatch
PASS window.event is set to the current event, which is the event passed to dispatch
